Adding RocksDB tests to the test suite and fixing failing tests

+++ src/rocksdb/build-rocksdb.sh
@@ -32,6 +32,13 @@ Options:
     Specify to use the ld.gold linker.
   --build-all-targets
     Specify to build tests, tools, and benchmarks
+  --target <target>
+    Explicitly specify an additional Makefile target to build. This option can be repeated and all
+    targets specified this way will be built.
+  --skip-link-dir-creation
+    Skip creation of a symlink directory tree for an out-of-source RocksDB build. This can be used
+    if we know this has already been done (e.g. when building a test right after building the
+    RocksDB library).
 EOT
 }
 
@@ -46,8 +53,10 @@ use_ld_gold=false
 extra_ldflags=""
 verbose=false
 build_all_targets=false
+skip_link_dir_creation=false
 
 make_opts=()
+make_targets=()
 
 while [ $# -ne 0 ]; do
   case "$1" in
@@ -95,6 +104,13 @@ while [ $# -ne 0 ]; do
     --build-all-targets)
       build_all_targets=true
     ;;
+    --target)
+      make_targets+=( "$2" )
+      shift
+    ;;
+    --skip-link-dir-creation)
+      skip_link_dir_creation=true
+    ;;
     *)
       print_help >&2
       echo >&2
@@ -140,7 +156,6 @@ if $use_ld_gold; then
   extra_ldflags="-fuse-ld=gold"
 fi
 
-make_targets=()
 case "$link_mode" in
   d)
     make_targets+=( shared_lib )
@@ -156,7 +171,9 @@ esac
 
 if [ "$debug_level" -gt 0 ] && $build_all_targets; then
   # We can only build tests if NDEBUG is not defined (otherwise e.g. db_test.cc fails to build).
-  make_targets+=( tests )
+  # TODO: try to build as many tests as possible in release mode.
+  # env_mirror_test does not seem to be built as part of the tests target, so we add it explicitly.
+  make_targets+=( tests env_mirror_test )
 fi
 
 if $build_all_targets; then
@@ -224,43 +241,47 @@ if [ -n "$extra_ldflags" ]; then
   make_opts+=( EXTRA_LDFLAGS="$extra_ldflags" )
 fi
 
-# -------------------------------------------------------------------------------------------------
-# Every command will be logged from this point on. All variables should have been set up above.
+if ! $skip_link_dir_creation; then
+  (
+    set -x
+    rm -rf "$link_dir"
 
-set -x
+    cd "$rocksdb_dir"
 
-rm -rf "$link_dir"
+    make clean  # ensure there are no build artifacts in the RocksDB source directory itself
 
-cd "$rocksdb_dir"
-make clean  # ensure there are no build artifacts in the RocksDB source directory itself
+    # Create a "link-only" directory inside of our build directory that mirrors the RocksDB source tree.
+    $CP -Rs "$rocksdb_dir" "$link_dir"
 
-# Create a "link-only" directory inside of our build directory that mirrors the RocksDB source tree.
-$CP -Rs "$rocksdb_dir" "$link_dir"
+    # Sync any new links that may have been created into our RocksDB build directory that may already
+    # contain build artifacts from an earlier build.
+    rsync -al "$link_dir/" "$rocksdb_build_dir"
 
-# Sync any new links that may have been created into our RocksDB build directory that may already
-# contain build artifacts from an earlier build.
-rsync -al "$link_dir/" "$rocksdb_build_dir"
+  )
+fi
 
+( set -x; mkdir -p "$rocksdb_build_dir" )
 cd "$rocksdb_build_dir"
 
 set +u  # make_opts may be empty and that we don't want that to be treated as an undefined variable
-make "${make_opts[@]}" "${make_targets[@]}"
 
-set +x -u
+( set -x; make "${make_opts[@]}" "${make_targets[@]}" )
+
+set -u
 
 # Now we're handling a weird issue that only happens during CLion-triggered builds.
-# The RocksDB library is somehow expected to exist in __default__/rocksdb-build instead of
-# e.g. Debug/rocksdb-build, so we just symlink it there.
+# The RocksDB library and test binaries are expected to exist in __default__/rocksdb-build instead
+# of e.g. Debug/rocksdb-build, so we just symlink them there.
 
 if [[ "$build_dir" =~ /[.]CLion.*/ ]]; then
   if [ -d "$build_dir/../__default__" ]; then
-    mkdir -p "$build_dir/../__default__/rocksdb-build"
-    for rocksdb_lib_path in "$rocksdb_build_dir"/librocksdb*; do
-      if [ -f "$rocksdb_lib_path" ]; then
-        dest_path="$build_dir/../__default__/rocksdb-build/${rocksdb_lib_path##*/}"
+    ( set -x; mkdir -p "$build_dir/../__default__/rocksdb-build" )
+    for symlink_target in "$rocksdb_build_dir"/librocksdb* "$rocksdb_build_dir"/*_test; do
+      if [ -f "$symlink_target" ]; then
+        dest_path="$build_dir/../__default__/rocksdb-build/${symlink_target##*/}"
         if [ ! -f "$dest_path" ]; then
-          echo "Creating a symlink '$dest_path' to '${rocksdb_lib_path}' for CLion"
-          ln -s "$rocksdb_lib_path" "$dest_path"
+          echo "Creating a symlink '$dest_path' to '${symlink_target}' for CLion"
+          ( set -x; ln -s "$symlink_target" "$dest_path" )
         fi
       fi
     done
